requirements.txt
# Core Dependencies
openai>=1.0.0  # OpenRouter uses OpenAI-compatible API
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Local STT (Open Source)
faster-whisper>=0.10.0

# Web Framework
gradio>=4.0.0
flask>=3.0.0

# Image Processing
pillow>=10.0.0

# Database (Local Storage)
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# HTTP Client
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Utilities
click>=8.1.0  # CLI tool
rich>=13.0.0  # Terminal formatting
